backend/requirements.txt
Django>=5.1.1,<6.0
djangorestframework==3.15.1
psycopg2-binary==2.9.7
